+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix freezing when running actions.
 - Add action to clear REPL output. #78
+- Add action to refresh all namespaces. #80
 
 ## 1.3.1